add berry, berry firmness, evolution triggers and growth rates

+++ src/models/language.rs
@@ -1,1 +1,1 @@
-pub type LanguageId = u16;
+pub type LanguageId = u8;
